Refactored security module
